Store `dist` snapshots in Git
The new `MAINLINE` environment variable designates the active line of
Biome versions configured by the files in the `src` directory. This
allows us to maintain just a single set of configuration files while
preserving backwards compatibility in the npm package via the snapshots
of the old build artefacts.

If we ever need to make a change to an obsolete line of Biome versions,
we can copy-paste the JSON file snapshots back into the JSON files in
the `src` directory and update the `MAINLINE` environment variable
accordingly.